+++ src/api/query.rs
@@ -12,7 +12,7 @@ use super::store::{MemPortalStore, PortalStore};
 use super::{ClientInfo, DEFAULT_NAME};
 use crate::api::results::{DescribeResponse, QueryResponse, Response};
 use crate::error::{PgWireError, PgWireResult};
-use crate::messages::data::ParameterDescription;
+use crate::messages::data::{NoData, ParameterDescription};
 use crate::messages::extendedquery::{
     Bind, BindComplete, Close, CloseComplete, Describe, Execute, Parse, ParseComplete,
     Sync as PgSync, TARGET_TYPE_BYTE_PORTAL, TARGET_TYPE_BYTE_STATEMENT,
@@ -27,6 +27,9 @@ pub trait SimpleQueryHandler: Send + Sync {
     /// Executed on `Query` request arrived. This is how postgres respond to
     /// simple query. The default implementation calls `do_query` with the
     /// incoming query string.
+    ///
+    /// This handle checks empty query by default, if the query string is empty
+    /// or `;`, it returns `EmptyQueryResponse` and does not call `self.do_query`.
     async fn on_query<C>(&self, client: &mut C, query: Query) -> PgWireResult<()>
     where
         C: ClientInfo + Sink<PgWireBackendMessage> + Unpin + Send + Sync,
@@ -43,6 +46,11 @@ pub trait SimpleQueryHandler: Send + Sync {
             let resp = self.do_query(client, query.query()).await?;
             for r in resp {
                 match r {
+                    Response::EmptyQuery => {
+                        client
+                            .feed(PgWireBackendMessage::EmptyQueryResponse(EmptyQueryResponse))
+                            .await?;
+                    }
                     Response::Query(results) => {
                         send_query_response(client, results, true).await?;
                     }
@@ -90,6 +98,10 @@ pub trait ExtendedQueryHandler: Send + Sync {
     /// Get a reference to associated `QueryParser` implementation
     fn query_parser(&self) -> Arc<Self::QueryParser>;
 
+    /// Called when client sends `parse` command.
+    ///
+    /// The default implementation parsed query with `Self::QueryParser` and
+    /// stores it in `Self::PortalStore`.
     async fn on_parse<C>(&self, client: &mut C, message: Parse) -> PgWireResult<()>
     where
         C: ClientInfo + Sink<PgWireBackendMessage> + Unpin + Send + Sync,
@@ -105,6 +117,10 @@ pub trait ExtendedQueryHandler: Send + Sync {
         Ok(())
     }
 
+    /// Called when client sends `bind` command.
+    ///
+    /// The default implementation associate parameters with previous parsed
+    /// statement and stores in `Self::PortalStore` as well.
     async fn on_bind<C>(&self, client: &mut C, message: Bind) -> PgWireResult<()>
     where
         C: ClientInfo + Sink<PgWireBackendMessage> + Unpin + Send + Sync,
@@ -125,6 +141,14 @@ pub trait ExtendedQueryHandler: Send + Sync {
         }
     }
 
+    /// Called when client sends `execute` command.
+    ///
+    /// The default implementation delegates the query to `self::do_query` and
+    /// sends response messages according to `Response` from `self::do_query`.
+    ///
+    /// Note that, different from `SimpleQueryHandler`, this implementation
+    /// won't check empty query because it cannot understand parsed
+    /// `Self::Statement`.
     async fn on_execute<C>(&self, client: &mut C, message: Execute) -> PgWireResult<()>
     where
         C: ClientInfo + Sink<PgWireBackendMessage> + Unpin + Send + Sync,
@@ -137,6 +161,12 @@ pub trait ExtendedQueryHandler: Send + Sync {
                 .do_query(client, portal.as_ref(), *message.max_rows() as usize)
                 .await?
             {
+                Response::EmptyQuery => {
+                    client.feed(PgWireBackendMessage::NoData(NoData)).await?;
+                    client
+                        .feed(PgWireBackendMessage::EmptyQueryResponse(EmptyQueryResponse))
+                        .await?;
+                }
                 Response::Query(results) => {
                     send_query_response(client, results, false).await?;
                 }
@@ -161,6 +191,9 @@ pub trait ExtendedQueryHandler: Send + Sync {
         }
     }
 
+    /// Called when client sends `describe` command.
+    ///
+    /// The default implementation delegates the call to `self::do_describe`.
     async fn on_describe<C>(&self, client: &mut C, message: Describe) -> PgWireResult<()>
     where
         C: ClientInfo + Sink<PgWireBackendMessage> + Unpin + Send + Sync,
@@ -217,6 +250,9 @@ pub trait ExtendedQueryHandler: Send + Sync {
         Ok(())
     }
 
+    /// Called when client sends `sync` command.
+    ///
+    /// The default implementation flushes client buffer.
     async fn on_sync<C>(&self, client: &mut C, _message: PgSync) -> PgWireResult<()>
     where
         C: ClientInfo + Sink<PgWireBackendMessage> + Unpin + Send + Sync,
@@ -227,6 +263,9 @@ pub trait ExtendedQueryHandler: Send + Sync {
         Ok(())
     }
 
+    /// Called when client sends `close` command.
+    ///
+    /// The default implementation closes certain statement or portal.
     async fn on_close<C>(&self, client: &mut C, message: Close) -> PgWireResult<()>
     where
         C: ClientInfo + Sink<PgWireBackendMessage> + Unpin + Send + Sync,